+++ renderer/primitives/Dimensions.h
@@ -1,14 +1,12 @@
 #ifndef SDL_PRIMITIVES_DIMENSIONS_H_
 #define SDL_PRIMITIVES_DIMENSIONS_H_
 
-#include <cstdint>
-
 struct Dimensions {
 	Dimensions();
-	Dimensions(int32_t width, int32_t height);
+	Dimensions(float width, float height);
 
-	int32_t w;
-	int32_t h;
+	float w;
+	float h;
 
 	static const Dimensions UNDEFINED;
 };
